urxvt, xorg improved. switched to agnoster
Turns out ttf-hack is already powerline compatible, for the most
part. I'll likely remove the missing symbols when I write my own
zsh theme (so far, only noticed the red X missing when a command
returns an error code; would be better to print the error code on
the right of the terminal).